+++ Source/FirstParty/Src/gui/Menu.hpp
@@ -32,8 +32,8 @@ public:
 
     virtual ~Menu();
 
-    void update(sf::RenderWindow& screen, const float time);
-    void update(sf::RenderWindow& screen, const float time, const MenuElementType::Type type);
+    void update(sf::RenderWindow& screen, const double time);
+    void update(sf::RenderWindow& screen, const double time, const MenuElementType::Type type);
 
     void registerOnClick(std::function<void(const Button& sender)> callback);
 
@@ -58,7 +58,7 @@ protected:
 private:
     virtual void drawAdditionalBackground(const DrawParameter& params);
     virtual void drawAdditionalForeground(const DrawParameter& params);
-    virtual void updated(sf::RenderWindow& screen, const float time);
+    virtual void updated(sf::RenderWindow& screen, const double time);
 
     template<class T>
     T* find(int id, const MenuElementType::Type type) const;
