+++ Source/FirstParty/Src/gui/MenuPanel.cpp
@@ -49,13 +49,13 @@ void MenuPanel::doDraw(const DrawParameter& params)
         (*it)->draw(params);
 }
 
-void MenuPanel::update(const sf::RenderWindow& screen, const float time, const sf::Vector2i& mouseOffset)
+void MenuPanel::update(const sf::RenderWindow& screen, const double time, const sf::Vector2i& mouseOffset)
 {
     for(auto it = begin(m_elements); it != end(m_elements); ++it)
         (*it)->update(screen, time, mouseOffset);
 }
 
-void MenuPanel::update(const sf::RenderWindow& screen, const float time, const MenuElementType::Type type, const sf::Vector2i& mouseOffset)
+void MenuPanel::update(const sf::RenderWindow& screen, const double time, const MenuElementType::Type type, const sf::Vector2i& mouseOffset)
 {
     for(auto it = begin(m_elements); it != end(m_elements); ++it)
     {
