+++ Source/FirstParty/Src/gui/MenuSprite.hpp
@@ -26,7 +26,7 @@ protected:
 
 private:
     void doDraw(const DrawParameter& params) override;
-    void updated(const sf::RenderWindow& screen, const float time, const sf::Vector2i& offset = sf::Vector2i(0, 0)) override;
+    void updated(const sf::RenderWindow& screen, const double time, const sf::Vector2i& offset = sf::Vector2i(0, 0)) override;
     void layoutUpdated(const sf::Vector2f& screenSize) override;
     void onDrawAdditionalForeground(const DrawParameter& params) override;
